Extract DeferredIntentConfirmationInterceptor.kt
