+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/intent/IntentConfirmationInterceptor.kt
@@ -2,7 +2,6 @@ package com.stripe.android.paymentelement.confirmation.intent
 
 import com.stripe.android.ConfirmStripeIntentParamsFactory
 import com.stripe.android.SharedPaymentTokenSessionPreview
-import com.stripe.android.common.exception.stripeErrorMessage
 import com.stripe.android.core.exception.StripeException
 import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
@@ -11,17 +10,13 @@ import com.stripe.android.core.strings.resolvableString
 import com.stripe.android.model.ConfirmPaymentIntentParams
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.PaymentMethod
-import com.stripe.android.model.PaymentMethodCode
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PaymentMethodExtraParams
 import com.stripe.android.model.PaymentMethodOptionsParams
 import com.stripe.android.model.RadarOptions
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.StripeIntent
-import com.stripe.android.model.setupFutureUsage
-import com.stripe.android.model.updateSetupFutureUsageWithPmoSfu
 import com.stripe.android.networking.StripeRepository
-import com.stripe.android.paymentelement.PaymentMethodOptionsSetupFutureUsagePreview
 import com.stripe.android.paymentelement.PreparePaymentMethodHandler
 import com.stripe.android.paymentelement.confirmation.ALLOWS_MANUAL_CONFIRMATION
 import com.stripe.android.paymentelement.confirmation.ConfirmationDefinition
@@ -29,18 +24,11 @@ import com.stripe.android.paymentelement.confirmation.ConfirmationHandler
 import com.stripe.android.paymentelement.confirmation.intent.IntentConfirmationDefinition.Args
 import com.stripe.android.payments.core.analytics.ErrorReporter
 import com.stripe.android.paymentsheet.CreateIntentCallback
-import com.stripe.android.paymentsheet.CreateIntentResult
-import com.stripe.android.paymentsheet.DeferredIntentValidator
 import com.stripe.android.paymentsheet.PaymentSheet
-import com.stripe.android.paymentsheet.R
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
-import com.stripe.android.utils.hasIntentToSetup
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.withTimeoutOrNull
 import javax.inject.Inject
 import javax.inject.Named
 import javax.inject.Provider
-import kotlin.time.Duration.Companion.seconds
 import com.stripe.android.R as PaymentsCoreR
 
 internal interface IntentConfirmationInterceptor {
@@ -156,46 +144,21 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
                         paymentMethodCreateParams = paymentMethodCreateParams,
                     )
                 } else {
-                    val intentConfiguration = initializationMode.intentConfiguration
-                    val productUsage = buildSet {
-                        addAll(paymentMethodCreateParams.attribution)
-                        add("deferred-intent")
-                        if (intentConfiguration.paymentMethodTypes.isEmpty()) {
-                            add("autopm")
-                        }
-                    }
-
-                    val params = paymentMethodCreateParams.copy(
-                        productUsage = productUsage,
-                    )
-
-                    return createPaymentMethod(params).fold(
-                        onSuccess = { paymentMethod ->
-                            handleDeferredIntent(
-                                intent = intent,
-                                intentConfiguration = intentConfiguration,
-                                paymentMethod = paymentMethod,
-                                paymentMethodOptionsParams = updatePaymentMethodOptionsParams(
-                                    code = paymentMethodCreateParams.typeCode,
-                                    intentConfiguration = initializationMode.intentConfiguration,
-                                    paymentMethodOptionsParams = paymentMethodOptionsParams
-                                ),
-                                paymentMethodExtraParams = paymentMethodExtraParams,
-                                shippingValues = shippingValues,
-                                shouldSavePaymentMethod = customerRequestedSave || shouldSavePaymentMethod(
-                                    paymentMethodOptionsParams = paymentMethodOptionsParams,
-                                    intentConfiguration = intentConfiguration
-                                ),
-                                hCaptchaToken = null
-                            )
-                        },
-                        onFailure = { error ->
-                            ConfirmationDefinition.Action.Fail(
-                                cause = error,
-                                message = error.stripeErrorMessage(),
-                                errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-                            )
-                        }
+                    DeferredIntentConfirmationInterceptor(
+                        stripeRepository = stripeRepository,
+                        errorReporter = errorReporter,
+                        intentCreationCallbackProvider = intentCreationCallbackProvider,
+                        allowsManualConfirmation = allowsManualConfirmation,
+                        publishableKeyProvider = publishableKeyProvider,
+                        stripeAccountIdProvider = stripeAccountIdProvider,
+                    ).handleNewPaymentMethod(
+                        intentConfiguration = initializationMode.intentConfiguration,
+                        intent = intent,
+                        paymentMethodCreateParams = paymentMethodCreateParams,
+                        paymentMethodOptionsParams = paymentMethodOptionsParams,
+                        paymentMethodExtraParams = paymentMethodExtraParams,
+                        shippingValues = shippingValues,
+                        customerRequestedSave = customerRequestedSave,
                     )
                 }
             }
@@ -249,23 +212,21 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
                         shippingValues = shippingValues,
                     )
                 } else {
-                    val updatedPaymentMethodOptionsParams = updatePaymentMethodOptionsParams(
-                        code = paymentMethod.type?.code,
+                    DeferredIntentConfirmationInterceptor(
+                        stripeRepository = stripeRepository,
+                        errorReporter = errorReporter,
+                        intentCreationCallbackProvider = intentCreationCallbackProvider,
+                        allowsManualConfirmation = allowsManualConfirmation,
+                        publishableKeyProvider = publishableKeyProvider,
+                        stripeAccountIdProvider = stripeAccountIdProvider,
+                    ).handleSavedPaymentMethod(
                         intentConfiguration = initializationMode.intentConfiguration,
-                        paymentMethodOptionsParams = paymentMethodOptionsParams
-                    )
-                    handleDeferredIntent(
                         intent = intent,
-                        intentConfiguration = initializationMode.intentConfiguration,
                         paymentMethod = paymentMethod,
-                        paymentMethodOptionsParams = updatedPaymentMethodOptionsParams,
+                        paymentMethodOptionsParams = paymentMethodOptionsParams,
                         paymentMethodExtraParams = paymentMethodExtraParams,
                         shippingValues = shippingValues,
-                        shouldSavePaymentMethod = shouldSavePaymentMethod(
-                            paymentMethodOptionsParams = updatedPaymentMethodOptionsParams,
-                            intentConfiguration = initializationMode.intentConfiguration
-                        ),
-                        hCaptchaToken = hCaptchaToken
+                        hCaptchaToken = hCaptchaToken,
                     )
                 }
             }
@@ -298,207 +259,6 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
                 )
             }
         }
-    }
-
-    private suspend fun handleDeferredIntent(
-        intent: StripeIntent,
-        intentConfiguration: PaymentSheet.IntentConfiguration,
-        paymentMethod: PaymentMethod,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
-        paymentMethodExtraParams: PaymentMethodExtraParams?,
-        shippingValues: ConfirmPaymentIntentParams.Shipping?,
-        shouldSavePaymentMethod: Boolean,
-        hCaptchaToken: String?
-    ): ConfirmationDefinition.Action<Args> {
-        return when (val callback = waitForIntentCallback()) {
-            is CreateIntentCallback -> {
-                handleDeferredIntentCreationFromPaymentMethod(
-                    intent = intent,
-                    createIntentCallback = callback,
-                    intentConfiguration = intentConfiguration,
-                    paymentMethod = paymentMethod,
-                    paymentMethodOptionsParams = paymentMethodOptionsParams,
-                    paymentMethodExtraParams = paymentMethodExtraParams,
-                    shouldSavePaymentMethod = shouldSavePaymentMethod,
-                    shippingValues = shippingValues,
-                    hCaptchaToken = hCaptchaToken
-                )
-            }
-
-            else -> {
-                val error = "${CreateIntentCallback::class.java.simpleName} must be implemented " +
-                    "when using IntentConfiguration with PaymentSheet"
-
-                errorReporter.report(ErrorReporter.ExpectedErrorEvent.CREATE_INTENT_CALLBACK_NULL)
-
-                ConfirmationDefinition.Action.Fail(
-                    cause = IllegalStateException(error),
-                    message = if (requestOptions.apiKeyIsLiveMode) {
-                        PaymentsCoreR.string.stripe_internal_error.resolvableString
-                    } else {
-                        error.resolvableString
-                    },
-                    errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-                )
-            }
-        }
-    }
-
-
-    private suspend fun createPaymentMethod(
-        params: PaymentMethodCreateParams,
-    ): Result<PaymentMethod> {
-        return stripeRepository.createPaymentMethod(
-            paymentMethodCreateParams = params,
-            options = requestOptions,
-        )
-    }
-
-    private suspend fun waitForIntentCallback(): CreateIntentCallback? {
-        return retrieveCallback() ?: run {
-            val callback = withTimeoutOrNull(PROVIDER_FETCH_TIMEOUT.seconds) {
-                var intentCallback: CreateIntentCallback? = null
-
-                while (intentCallback == null) {
-                    delay(PROVIDER_FETCH_INTERVAL)
-                    intentCallback = retrieveCallback()
-                }
-
-                intentCallback
-            }
-
-            if (callback != null) {
-                errorReporter.report(ErrorReporter.SuccessEvent.FOUND_CREATE_INTENT_CALLBACK_WHILE_POLLING)
-            }
-
-            callback
-        }
-    }
-
-    private fun retrieveCallback(): CreateIntentCallback? {
-        return intentCreationCallbackProvider.get()
-    }
-
-    private suspend fun handleDeferredIntentCreationFromPaymentMethod(
-        intent: StripeIntent,
-        createIntentCallback: CreateIntentCallback,
-        intentConfiguration: PaymentSheet.IntentConfiguration,
-        paymentMethod: PaymentMethod,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
-        paymentMethodExtraParams: PaymentMethodExtraParams?,
-        shouldSavePaymentMethod: Boolean,
-        shippingValues: ConfirmPaymentIntentParams.Shipping?,
-        hCaptchaToken: String?
-    ): ConfirmationDefinition.Action<Args> {
-        val result = createIntentCallback.onCreateIntent(
-            paymentMethod = paymentMethod,
-            shouldSavePaymentMethod = shouldSavePaymentMethod,
-        )
-
-        return when (result) {
-            is CreateIntentResult.Success -> {
-                if (result.clientSecret == IntentConfirmationInterceptor.COMPLETE_WITHOUT_CONFIRMING_INTENT) {
-                    ConfirmationDefinition.Action.Complete(
-                        intent = intent,
-                        deferredIntentConfirmationType = DeferredIntentConfirmationType.None,
-                        completedFullPaymentFlow = true,
-                    )
-                } else {
-                    handleDeferredIntentCreationSuccess(
-                        clientSecret = result.clientSecret,
-                        intentConfiguration = intentConfiguration,
-                        paymentMethod = paymentMethod,
-                        paymentMethodOptionsParams = paymentMethodOptionsParams,
-                        paymentMethodExtraParams = paymentMethodExtraParams,
-                        shippingValues = shippingValues,
-                        hCaptchaToken = hCaptchaToken
-                    )
-                }
-            }
-
-            is CreateIntentResult.Failure -> {
-                val exception = CreateIntentCallbackFailureException(result.cause)
-                ConfirmationDefinition.Action.Fail(
-                    cause = exception,
-                    message = result.displayMessage?.resolvableString
-                        ?: exception.stripeErrorMessage(),
-                    errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-                )
-            }
-        }
-    }
-
-    private suspend fun handleDeferredIntentCreationSuccess(
-        clientSecret: String,
-        intentConfiguration: PaymentSheet.IntentConfiguration,
-        paymentMethod: PaymentMethod,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
-        paymentMethodExtraParams: PaymentMethodExtraParams?,
-        shippingValues: ConfirmPaymentIntentParams.Shipping?,
-        hCaptchaToken: String?
-    ): ConfirmationDefinition.Action<Args> {
-        return retrieveStripeIntent(clientSecret).mapCatching { intent ->
-            if (intent.isConfirmed) {
-                failIfSetAsDefaultFeatureIsEnabled(paymentMethodExtraParams)
-
-                ConfirmationDefinition.Action.Complete(
-                    intent = intent,
-                    deferredIntentConfirmationType = DeferredIntentConfirmationType.Server,
-                    completedFullPaymentFlow = true,
-                )
-            } else if (intent.requiresAction()) {
-                createHandleNextActionStep(clientSecret, intent, paymentMethod)
-            } else {
-                DeferredIntentValidator.validate(intent, intentConfiguration, allowsManualConfirmation, paymentMethod)
-                createConfirmStep(
-                    clientSecret,
-                    intent,
-                    shippingValues,
-                    paymentMethod,
-                    paymentMethodOptionsParams = paymentMethodOptionsParams,
-                    paymentMethodExtraParams = paymentMethodExtraParams,
-                    isDeferred = true,
-                    intentConfigSetupFutureUsage = intentConfiguration
-                        .mode.setupFutureUse?.toConfirmParamsSetupFutureUsage(),
-                    hCaptchaToken = hCaptchaToken
-                )
-            }
-        }.getOrElse { error ->
-            ConfirmationDefinition.Action.Fail(
-                cause = error,
-                message = error.stripeErrorMessage(),
-                errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-            )
-        }
-    }
-
-    private fun createHandleNextActionStep(
-        clientSecret: String,
-        intent: StripeIntent,
-        paymentMethod: PaymentMethod
-    ): ConfirmationDefinition.Action<Args> {
-        return runCatching<ConfirmationDefinition.Action<Args>> {
-            DeferredIntentValidator.validatePaymentMethod(intent, paymentMethod)
-
-            ConfirmationDefinition.Action.Launch(
-                launcherArguments = Args.NextAction(clientSecret),
-                deferredIntentConfirmationType = DeferredIntentConfirmationType.Server,
-                receivesResultInProcess = false,
-            )
-        }.getOrElse {
-            ConfirmationDefinition.Action.Fail(
-                cause = InvalidDeferredIntentUsageException(),
-                message = resolvableString(R.string.stripe_paymentsheet_invalid_deferred_intent_usage),
-                errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
-            )
-        }
-    }
-
-    private suspend fun retrieveStripeIntent(clientSecret: String): Result<StripeIntent> {
-        return stripeRepository.retrieveStripeIntent(
-            clientSecret = clientSecret,
-            options = requestOptions,
-        )
     }
 
     private fun createConfirmStep(
@@ -535,7 +295,6 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
             receivesResultInProcess = false,
         )
     }
-
     private fun createConfirmStep(
         clientSecret: String,
         intent: StripeIntent,
@@ -581,81 +340,6 @@ internal class DefaultIntentConfirmationInterceptor @Inject constructor(
             errorType = ConfirmationHandler.Result.Failed.ErrorType.Payment,
         )
     }
-
-    private fun failIfSetAsDefaultFeatureIsEnabled(paymentMethodExtraParams: PaymentMethodExtraParams?) {
-        // Ideally, we would crash anytime the set as default checkbox is shown, rather than just when it is checked.
-        // We could check if it is shown by asserting that setAsDefault != null instead of asserting that it is true.
-        // However, we don't have good end-to-end test coverage of this for now, so if we made a change to start
-        // sending the set as default flag as false more frequently, we could accidentally start failing here more
-        // often as well.
-        val setAsDefaultChecked = when (paymentMethodExtraParams) {
-            is PaymentMethodExtraParams.Card -> paymentMethodExtraParams.setAsDefault == true
-            is PaymentMethodExtraParams.USBankAccount -> paymentMethodExtraParams.setAsDefault == true
-            is PaymentMethodExtraParams.Link -> paymentMethodExtraParams.setAsDefault == true
-            is PaymentMethodExtraParams.SepaDebit -> paymentMethodExtraParams.setAsDefault == true
-            is PaymentMethodExtraParams.BacsDebit, null -> false
-        }
-
-        if (setAsDefaultChecked && !requestOptions.apiKeyIsLiveMode) {
-            throw IllegalStateException(
-                "(Test-mode only error) The default payment methods feature is not yet supported with deferred " +
-                    "server-side confirmation. Please contact us if you'd like to use this feature via a Github " +
-                    "issue on stripe-android."
-            )
-        }
-    }
-
-    private fun PaymentSheet.IntentConfiguration.SetupFutureUse.toConfirmParamsSetupFutureUsage():
-        ConfirmPaymentIntentParams.SetupFutureUsage {
-        return when (this) {
-            PaymentSheet.IntentConfiguration.SetupFutureUse.OffSession -> {
-                ConfirmPaymentIntentParams.SetupFutureUsage.OffSession
-            }
-            PaymentSheet.IntentConfiguration.SetupFutureUse.OnSession -> {
-                ConfirmPaymentIntentParams.SetupFutureUsage.OnSession
-            }
-            PaymentSheet.IntentConfiguration.SetupFutureUse.None -> {
-                ConfirmPaymentIntentParams.SetupFutureUsage.None
-            }
-        }
-    }
-
-    /**
-     * [PaymentSheet.IntentConfiguration.Mode.Payment.PaymentMethodOptions] does not require setting PMO SFU on the
-     * intent. If PMO SFU value exists in the configuration, set it in the PaymentMethodOptionsParams.
-     */
-    @OptIn(PaymentMethodOptionsSetupFutureUsagePreview::class)
-    private fun updatePaymentMethodOptionsParams(
-        code: PaymentMethodCode?,
-        intentConfiguration: PaymentSheet.IntentConfiguration,
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?
-    ): PaymentMethodOptionsParams? {
-        val paymentMethodType = PaymentMethod.Type.fromCode(code) ?: return paymentMethodOptionsParams
-        return (intentConfiguration.mode as? PaymentSheet.IntentConfiguration.Mode.Payment)
-            ?.paymentMethodOptions
-            ?.setupFutureUsageValues?.let { values ->
-                values[paymentMethodType]?.toConfirmParamsSetupFutureUsage()?.let { configPmoSfu ->
-                    if (paymentMethodOptionsParams != null) {
-                        paymentMethodOptionsParams.updateSetupFutureUsageWithPmoSfu(configPmoSfu)
-                    } else {
-                        PaymentMethodOptionsParams.SetupFutureUsage(
-                            paymentMethodType = paymentMethodType,
-                            setupFutureUsage = configPmoSfu
-                        )
-                    }
-                }
-            } ?: paymentMethodOptionsParams
-    }
-
-    private fun shouldSavePaymentMethod(
-        paymentMethodOptionsParams: PaymentMethodOptionsParams?,
-        intentConfiguration: PaymentSheet.IntentConfiguration
-    ): Boolean {
-        return paymentMethodOptionsParams?.setupFutureUsage()?.hasIntentToSetup() == true ||
-            (intentConfiguration.mode as? PaymentSheet.IntentConfiguration.Mode.Payment)
-                ?.setupFutureUse?.toConfirmParamsSetupFutureUsage()?.hasIntentToSetup() == true
-    }
-
     internal companion object {
         const val PROVIDER_FETCH_TIMEOUT = 2
         const val PROVIDER_FETCH_INTERVAL = 5L
